requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
mysql-connector-python
python-dotenv
databases[mysql]
cryptography
pytest
faker # to generate fake data 